Move logger to `utils.logging` module
