+++ py_neuromodulation/__init__.py
@@ -2,7 +2,6 @@ import os
 import platform
 from pathlib import PurePath
 from importlib.metadata import version
-from py_neuromodulation.utils.logging import NMLogger
 
 #####################################
 # Globals and environment variables #
@@ -56,12 +55,6 @@ if KEY in LSL_DICT:
 # Create global dictionary to store user-defined features
 user_features = {}
 
-
-######################################
-# Logger initialization and settings #
-######################################
-
-logger = NMLogger(__name__)  # logger initialization first to prevent circular import
 
 ####################################
 # API: Exposed classes and methods #
