+++ py_neuromodulation/utils/__init__.py
@@ -1,2 +1,5 @@
 from .channels import *
+from .io import *
 from . import types
+
+from .logging import logger
